@@ -1,15 +1,12 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
-
 
 import { AppComponent } from './app.component';
 import { GamesComponent } from './games/games.component';
 import { GameDetailsComponent } from './game-details/game-details.component';
-
-
 
 let routes = [{
     path: 'games',
@@ -32,7 +29,7 @@ const rootRouting: ModuleWithProviders = RouterModule.forRoot(routes, { useHash:
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     rootRouting
   ],
   providers: [],
